maj resolce conflict dolibarr 3.5.0

+++ htdocs/compta/tva/fiche.php
@@ -1,6 +1,6 @@
 <?php
 /* Copyright (C) 2003      Rodolphe Quiedeville <rodolphe@quiedeville.org>
- * Copyright (C) 2004-2010 Laurent Destailleur  <eldy@users.sourceforge.net>
+ * Copyright (C) 2004-2013 Laurent Destailleur  <eldy@users.sourceforge.net>
  * Copyright (C) 2005-2013 Regis Houssin        <regis.houssin@capnetworks.com>
  *
  * This program is free software; you can redistribute it and/or modify
@@ -31,10 +31,8 @@ $langs->load("compta");
 $langs->load("banks");
 $langs->load("bills");
 
-$id=GETPOST("id");
+$id=GETPOST("id",'int');
 $action=GETPOST('action');
-
-$mesg = '';
 
 // Security check
 $socid = isset($_GET["socid"])?$_GET["socid"]:'';
@@ -82,14 +80,14 @@ if ($action == 'add' && $_POST["cancel"] <> $langs->trans("Cancel"))
     else
     {
         $db->rollback();
-        $mesg='<div class="error">'.$tva->error.'</div>';
+        setEventMessage($tva->error, 'errors');
         $action="create";
     }
 }
 
 if ($action == 'delete')
 {
-    $result=$tva->fetch($_GET['id']);
+    $result=$tva->fetch($id);
 
 	if ($tva->rappro == 0)
 	{
@@ -102,10 +100,10 @@ if ($action == 'delete')
 			{
 				$accountline=new AccountLine($db);
 				$result=$accountline->fetch($tva->fk_bank);
-				$result=$accountline->delete($user);
+				if ($result > 0) $result=$accountline->delete($user);	// $result may be 0 if not found (when bank entry was deleted manually and fk_bank point to nothing)
 			}
 
-			if ($result > 0)
+			if ($result >= 0)
 			{
 				$db->commit();
 				header("Location: ".DOL_URL_ROOT.'/compta/tva/reglement.php');
@@ -115,18 +113,18 @@ if ($action == 'delete')
 			{
 				$tva->error=$accountline->error;
 				$db->rollback();
-				$mesg='<div class="error">'.$tva->error.'</div>';
+				setEventMessage($tva->error,'errors');
 			}
 	    }
 	    else
 	    {
 	        $db->rollback();
-	        $mesg='<div class="error">'.$tva->error.'</div>';
+	        setEventMessage($tva->error,'errors');
 	    }
 	}
 	else
 	{
-        $mesg='<div class="error">Error try do delete a line linked to a conciliated bank transaction</div>';
+        setEventMessage('Error try do delete a line linked to a conciliated bank transaction','errors');
 	}
 }
 
@@ -158,8 +156,6 @@ if ($action == 'create')
     print '<input type="hidden" name="action" value="add">';
 
     print_fiche_titre($langs->trans("NewVATPayment"));
-
-    if ($mesg) print $mesg;
 
     print '<table class="border" width="100%">';
 
@@ -213,8 +209,6 @@ if ($action == 'create')
 
 if ($id)
 {
-    if ($mesg) print $mesg;
-
 	$h = 0;
 	$head[$h][0] = DOL_URL_ROOT.'/compta/tva/fiche.php?id='.$vatpayment->id;
 	$head[$h][1] = $langs->trans('Card');
@@ -274,9 +268,20 @@ if ($id)
 	*/
 	print "<div class=\"tabsAction\">\n";
 	if ($vatpayment->rappro == 0)
-		print '<a class="butActionDelete" href="fiche.php?id='.$vatpayment->id.'&action=delete">'.$langs->trans("Delete").'</a>';
+	{
+		if (! empty($user->rights->tax->charges->supprimer))
+		{
+			print '<a class="butActionDelete" href="fiche.php?id='.$vatpayment->id.'&action=delete">'.$langs->trans("Delete").'</a>';
+		}
+		else
+		{
+			print '<a class="butActionRefused" href="#" title="'.(dol_escape_htmltag($langs->trans("NotAllowed"))).'">'.$langs->trans("Delete").'</a>';
+		}
+	}
 	else
+	{
 		print '<a class="butActionRefused" href="#" title="'.$langs->trans("LinkedToAConcialitedTransaction").'">'.$langs->trans("Delete").'</a>';
+	}
 	print "</div>";
 }
